delete confirmation message and navbar user detail added

+++ src/pages/MyCart/MyCart.jsx
@@ -9,18 +9,32 @@ const MyCart = () => {
 
   const handleDelete = (_id) => {
     // console.log(_id);
-    fetch(`http://localhost:5000/mycart/${_id}`,{
-      method:'DELETE',
-    })
-    .then(res=>res.json())
-    .then(data=>{
-      if(data.deletedCount>0){
-        console.log('deleted successfully');
-        const remainingProducts = cartProducts.filter(cartProduct=>cartProduct._id !==_id)
-        setCartProducts(remainingProducts); 
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`http://localhost:5000/mycart/${_id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            if (data.deletedCount > 0) {
+              console.log("deleted successfully");
+              Swal.fire("Deleted!", "Your file has been deleted.", "success");
+              const remainingProducts = cartProducts.filter(
+                (cartProduct) => cartProduct._id !== _id
+              );
+              setCartProducts(remainingProducts);
+            }
+          });
       }
-    })
-    
+    });
   };
 
   return (
